+++ common/src/main/java/net/spell_engine/api/entity/EvasionLogic.java
@@ -8,6 +8,7 @@ import net.minecraft.sound.SoundEvents;
 import net.minecraft.util.math.Vec3d;
 import net.spell_engine.SpellEngineMod;
 import net.spell_engine.api.event.CombatEvents;
+import net.spell_engine.api.spell.fx.Animation;
 import net.spell_engine.api.spell.fx.Sound;
 import net.spell_engine.api.tags.SpellEngineDamageTypeTags;
 import net.spell_engine.fx.SpellEngineSounds;
@@ -56,11 +57,12 @@ public class EvasionLogic {
     }
 
     private static final Sound evadeSound = new Sound("spell_engine:dodge", 1.0F, 1.0F, 0.1F);
+    private static final Animation evadeAnimation = new Animation("spell_engine:dodge");
     public static void onEvade(LivingEntity entity, float damage, DamageSource source) {
         // System.out.println("SpellEngine: " + entity.getName().getString() + " evaded damage from " + source.getName() + "!");
         if (entity instanceof ServerPlayerEntity player) {
             var tracker = PlayerLookup.tracking(player);
-            AnimationHelper.sendAnimation(player, tracker, SpellCast.Animation.MISC, "spell_engine:dodge", 1F);
+            AnimationHelper.sendAnimation(player, tracker, SpellCast.Animation.MISC, evadeAnimation, 1F);
         }
         entity.getWorld().playSound(null, entity.getX(), entity.getY(), entity.getZ(), SpellEngineSounds.DODGE.soundEvent(), entity.getSoundCategory(), 1.0F, evadeSound.randomizedPitch());
         CombatEvents.ENTITY_EVASION.invoke(listener -> listener.onEntityEvasion(new CombatEvents.EntityEvasion.Args(entity, damage, source)));